+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## unreleased
+
+- Fix kotlin Duration is converted to NSNumber instead of long in bindings
+
 ## v0.1.3
 
 - Remove non nullable primitives conversion to and from NSNumber in bindings